+++ source/raw/bin/x64/plugins/cyber_engine_tweaks/mods/Character Preset Manager (CET)/modules/state.lua
@@ -30,6 +30,7 @@ state = {
     pendingOverwriteFingerprint = nil,
     pendingRemoveFolder = nil,
     selectedBundleFile = nil,
+    folderBundlePreview = nil,
     sortMode = "name",
   },
   load = {